fix(app): name a worktree session's project after the repository
A worktree session's path points inside the tree, so both the row label
and the project grouping took the last path segment — the worktree's own
name — and a session in calm-meadow read "calm-meadow · calm-meadow" and
would have grouped into a phantom "calm-meadow" project instead of its
repository's card.

The pure path helpers move to utils/worktreePath.ts (re-exported from
utils/worktree.ts) so sync/projectGroups can use getRepoPath without
pulling the machine RPC plumbing into the storage import cycle.
