+++ packages/happy-app/sources/utils/worktree.ts
@@ -4,12 +4,12 @@
 
 import { machineBash } from '@/sync/ops';
 import { appendWorktreeNameSuffix, normalizeWorktreeName } from './worktreeName';
+import { WORKTREE_DIR, WORKTREE_PATH_MARKER } from './worktreePath';
 
-/** Relative path prefix where worktrees are stored inside a repo */
-export const WORKTREE_DIR = '.dev/worktree';
+// Path helpers moved to worktreePath.ts (pure, importable from sync modules);
+// re-exported here so existing imports keep working.
+export * from './worktreePath';
 
-/** Absolute path marker used to detect worktree paths */
-export const WORKTREE_PATH_MARKER = `/${WORKTREE_DIR}/`;
 
 // --- Name generation ---
 
@@ -184,21 +184,3 @@ export async function removeWorktree(
     };
 }
 
-/** Check if a path is inside a worktree */
-export function isWorktreePath(path: string): boolean {
-    return path.includes(WORKTREE_PATH_MARKER);
-}
-
-/** Extract the main repository checkout path from a possibly-worktree path */
-export function getRepoPath(path: string): string {
-    const idx = path.indexOf(WORKTREE_PATH_MARKER);
-    if (idx === -1) return path;
-    return path.slice(0, idx);
-}
-
-/** Extract the worktree name from a worktree path, or null if not a worktree */
-export function getWorktreeName(path: string): string | null {
-    const idx = path.indexOf(WORKTREE_PATH_MARKER);
-    if (idx === -1) return null;
-    return path.slice(idx + WORKTREE_PATH_MARKER.length);
-}
