+++ packages/happy-app/sources/utils/worktreePath.ts
@@ -1,0 +1,30 @@
+/**
+ * Pure worktree path helpers. Split from worktree.ts on purpose: that module
+ * pulls in machine RPC plumbing (sync/ops), and these functions are needed by
+ * sync/projectGroups, which must stay free of the storage import cycle.
+ */
+
+/** Relative path prefix where worktrees are stored inside a repo */
+export const WORKTREE_DIR = '.dev/worktree';
+
+/** Absolute path marker used to detect worktree paths */
+export const WORKTREE_PATH_MARKER = `/${WORKTREE_DIR}/`;
+
+/** Check if a path is inside a worktree */
+export function isWorktreePath(path: string): boolean {
+    return path.includes(WORKTREE_PATH_MARKER);
+}
+
+/** Extract the main repository checkout path from a possibly-worktree path */
+export function getRepoPath(path: string): string {
+    const idx = path.indexOf(WORKTREE_PATH_MARKER);
+    if (idx === -1) return path;
+    return path.slice(0, idx);
+}
+
+/** Extract the worktree name from a worktree path, or null if not a worktree */
+export function getWorktreeName(path: string): string | null {
+    const idx = path.indexOf(WORKTREE_PATH_MARKER);
+    if (idx === -1) return null;
+    return path.slice(idx + WORKTREE_PATH_MARKER.length);
+}
